requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.0
redis[hiredis]==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-jaeger==1.21.0
python-json-logger==2.0.7
tenacity==8.2.3
opensearch-py==2.3.1
faiss-cpu==1.7.4
locust==2.17.0
aiocache==0.12.2
aioredis==2.0.1
pybreaker==0.8.2
structlog==23.2.0
openai==1.3.0